% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humor_dataset}
\alias{humor_dataset}
\title{A list of English words with the "humor ratings" attached.}
\format{
A data frame with 4997 rows and 16 variables:
\describe{
  \item{word}{string of the actual word}
  \item{mean}{mean of humor rating across all audiences}
  \item{mean_F}{mean of humor rating (women)}
  \item{mean_M}{mean of humor rating (men)}
  \item{mean_old}{mean of humor rating (old)}
  \item{mean_young}{mean of humor rating (young)}
  \item{n}{audience size}
  \item{n_F}{audience size (women)}
  \item{n_M}{audience size (men)}
  \item{n_old}{audience size (old)}
  \item{n_young}{audience size (young)}
  \item{sd}{sd of humor rating across all audiences}
  \item{sd_F}{sd humor rating (women)}
  \item{sd_M}{sd of humor rating (men)}
  \item{sd_old}{sd humor rating (old)}
  \item{sd_young}{sd of humor rating (young)}

}
}
\source{
\url{https://github.com/tomasengelthaler/HumorNorms}
}
\usage{
data(humor_dataset)
}
\description{
A dataset compiled by Tomas Englethaler for his research on humor.
https://github.com/tomasengelthaler/HumorNorms
Please visit his page for more details on the methodology used to score words.
}
\keyword{datasets}
